requirements.txt
Flask==0.12.2
nltk
numpy
argparse
unidecode
textblob
python-rake